requirements.txt
flask
tensorflow
pillow
numpy
flask-sqlalchemy
flask-login
flask-wtf
werkzeug
